src/main/webapp/app/entities/references/delete/references-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import {ComponentFixture, fakeAsync, inject, TestBed, tick} from '@angular/core/testing';
import {HttpResponse} from '@angular/common/http';
import {HttpClientTestingModule} from '@angular/common/http/testing';
import {of} from 'rxjs';
import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';

import {ReferencesService} from '../service/references.service';

import {ReferencesDeleteDialogComponent} from './references-delete-dialog.component';

describe('References Management Delete Component', () => {
  let comp: ReferencesDeleteDialogComponent;
  let fixture: ComponentFixture<ReferencesDeleteDialogComponent>;
  let service: ReferencesService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [ReferencesDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(ReferencesDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ReferencesDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ReferencesService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
